requirements.txt
django==1.10.5
django-uuidfield
django-multiselectfield
djangorestframework
django-cors-headers
djangorestframework-jwt
biopython
django-filter
django-stored-messages
bcbio-gff
requests
